+++ tools4a-fragment-ktx/src/androidTest/java/com/github/panpf/tools4a/fragment/ktx/test/FragmentxTest.kt
@@ -29,10 +29,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.viewpager.widget.ViewPager
 import com.github.panpf.tools4a.fragment.ktx.*
-import com.github.panpf.tools4a.test.ktx.getActivitySync
-import com.github.panpf.tools4a.test.ktx.getFragmentSync
-import com.github.panpf.tools4a.test.ktx.launchActivityWithUse
-import com.github.panpf.tools4a.test.ktx.launchFragmentInContainer
+import com.github.panpf.tools4a.test.ktx.*
 import com.github.panpf.tools4j.test.ktx.assertNoThrow
 import com.github.panpf.tools4j.test.ktx.assertThrow
 import org.junit.Assert.*
@@ -116,17 +113,15 @@ class FragmentxTest {
             var userVisibleChildViewPagerFragment = activity.findUserVisibleChildFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
             assertEquals(0, userVisibleChildViewPagerFragment.tabIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms() {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
             userVisibleChildViewPagerFragment = activity.findUserVisibleChildFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
             assertEquals(1, userVisibleChildViewPagerFragment.tabIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
             }
-            Thread.sleep(200)
             userVisibleChildViewPagerFragment = activity.findUserVisibleChildFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
             assertEquals(2, userVisibleChildViewPagerFragment.tabIndex)
         }
@@ -140,17 +135,15 @@ class FragmentxTest {
         var userVisibleChildViewPagerFragment = fragment.findUserVisibleChildFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
         assertEquals(0, userVisibleChildViewPagerFragment.tabIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms() {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment = fragment.findUserVisibleChildFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
         assertEquals(1, userVisibleChildViewPagerFragment.tabIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment = fragment.findUserVisibleChildFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
         assertEquals(2, userVisibleChildViewPagerFragment.tabIndex)
 
@@ -163,17 +156,15 @@ class FragmentxTest {
         var userVisibleChildViewPagerFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
         assertEquals(0, userVisibleChildViewPagerFragment1.tabIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
         assertEquals(1, userVisibleChildViewPagerFragment1.tabIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentByUserVisibleHint() as UserVisibleHintChildViewPagerFragment
         assertEquals(2, userVisibleChildViewPagerFragment1.tabIndex)
     }
@@ -191,26 +182,23 @@ class FragmentxTest {
             assertEquals(0, userVisibleChildTextFragment.tabIndex)
             assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
             userVisibleChildTextFragment = activity.findUserVisibleChildFragmentRecursivelyByUserVisibleHint() as TextFragment
             assertEquals(1, userVisibleChildTextFragment.tabIndex)
             assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
             }
-            Thread.sleep(200)
             userVisibleChildTextFragment = activity.findUserVisibleChildFragmentRecursivelyByUserVisibleHint() as TextFragment
             assertEquals(2, userVisibleChildTextFragment.tabIndex)
             assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 userVisibleChildTextFragment.parentFragment!!.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
             userVisibleChildTextFragment = activity.findUserVisibleChildFragmentRecursivelyByUserVisibleHint() as TextFragment
             assertEquals(2, userVisibleChildTextFragment.tabIndex)
             assertEquals(1, userVisibleChildTextFragment.pageIndex)
@@ -226,26 +214,23 @@ class FragmentxTest {
         assertEquals(0, userVisibleChildTextFragment.tabIndex)
         assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment = fragment.findUserVisibleChildFragmentRecursivelyByUserVisibleHint() as TextFragment
         assertEquals(1, userVisibleChildTextFragment.tabIndex)
         assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment = fragment.findUserVisibleChildFragmentRecursivelyByUserVisibleHint() as TextFragment
         assertEquals(2, userVisibleChildTextFragment.tabIndex)
         assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             userVisibleChildTextFragment.parentFragment!!.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment = fragment.findUserVisibleChildFragmentRecursivelyByUserVisibleHint() as TextFragment
         assertEquals(2, userVisibleChildTextFragment.tabIndex)
         assertEquals(1, userVisibleChildTextFragment.pageIndex)
@@ -260,26 +245,23 @@ class FragmentxTest {
         assertEquals(0, userVisibleChildTextFragment1.tabIndex)
         assertEquals(0, userVisibleChildTextFragment1.pageIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentRecursivelyByUserVisibleHint() as TextFragment
         assertEquals(1, userVisibleChildTextFragment1.tabIndex)
         assertEquals(0, userVisibleChildTextFragment1.pageIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentRecursivelyByUserVisibleHint() as TextFragment
         assertEquals(2, userVisibleChildTextFragment1.tabIndex)
         assertEquals(0, userVisibleChildTextFragment1.pageIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             userVisibleChildTextFragment1.parentFragment!!.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentRecursivelyByUserVisibleHint() as TextFragment
         assertEquals(2, userVisibleChildTextFragment1.tabIndex)
         assertEquals(1, userVisibleChildTextFragment1.pageIndex)
@@ -297,17 +279,15 @@ class FragmentxTest {
             var userVisibleChildViewPagerFragment = activity.findUserVisibleChildFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
             assertEquals(0, userVisibleChildViewPagerFragment.tabIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
             userVisibleChildViewPagerFragment = activity.findUserVisibleChildFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
             assertEquals(1, userVisibleChildViewPagerFragment.tabIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
             }
-            Thread.sleep(200)
             userVisibleChildViewPagerFragment = activity.findUserVisibleChildFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
             assertEquals(2, userVisibleChildViewPagerFragment.tabIndex)
         }
@@ -321,17 +301,15 @@ class FragmentxTest {
         var userVisibleChildViewPagerFragment = fragment.findUserVisibleChildFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
         assertEquals(0, userVisibleChildViewPagerFragment.tabIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment = fragment.findUserVisibleChildFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
         assertEquals(1, userVisibleChildViewPagerFragment.tabIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment = fragment.findUserVisibleChildFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
         assertEquals(2, userVisibleChildViewPagerFragment.tabIndex)
 
@@ -344,17 +322,15 @@ class FragmentxTest {
         var userVisibleChildViewPagerFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
         assertEquals(0, userVisibleChildViewPagerFragment1.tabIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
         assertEquals(1, userVisibleChildViewPagerFragment1.tabIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildViewPagerFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentByResumed() as ResumeOnlyCurrentChildViewPagerFragment
         assertEquals(2, userVisibleChildViewPagerFragment1.tabIndex)
     }
@@ -372,26 +348,23 @@ class FragmentxTest {
             assertEquals(0, userVisibleChildTextFragment.tabIndex)
             assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
             userVisibleChildTextFragment = activity.findUserVisibleChildFragmentRecursivelyByResumed() as TextFragment
             assertEquals(1, userVisibleChildTextFragment.tabIndex)
             assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 it.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
             }
-            Thread.sleep(200)
             userVisibleChildTextFragment = activity.findUserVisibleChildFragmentRecursivelyByResumed() as TextFragment
             assertEquals(2, userVisibleChildTextFragment.tabIndex)
             assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 userVisibleChildTextFragment.parentFragment!!.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
             userVisibleChildTextFragment = activity.findUserVisibleChildFragmentRecursivelyByResumed() as TextFragment
             assertEquals(2, userVisibleChildTextFragment.tabIndex)
             assertEquals(1, userVisibleChildTextFragment.pageIndex)
@@ -407,26 +380,23 @@ class FragmentxTest {
         assertEquals(0, userVisibleChildTextFragment.tabIndex)
         assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment = fragment.findUserVisibleChildFragmentRecursivelyByResumed() as TextFragment
         assertEquals(1, userVisibleChildTextFragment.tabIndex)
         assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment = fragment.findUserVisibleChildFragmentRecursivelyByResumed() as TextFragment
         assertEquals(2, userVisibleChildTextFragment.tabIndex)
         assertEquals(0, userVisibleChildTextFragment.pageIndex)
 
-        fragmentScenario.onFragment {
+        fragmentScenario.onFragmentSleep100ms {
             userVisibleChildTextFragment.parentFragment!!.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment = fragment.findUserVisibleChildFragmentRecursivelyByResumed() as TextFragment
         assertEquals(2, userVisibleChildTextFragment.tabIndex)
         assertEquals(1, userVisibleChildTextFragment.pageIndex)
@@ -441,26 +411,23 @@ class FragmentxTest {
         assertEquals(0, userVisibleChildTextFragment1.tabIndex)
         assertEquals(0, userVisibleChildTextFragment1.pageIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentRecursivelyByResumed() as TextFragment
         assertEquals(1, userVisibleChildTextFragment1.tabIndex)
         assertEquals(0, userVisibleChildTextFragment1.pageIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             it.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentRecursivelyByResumed() as TextFragment
         assertEquals(2, userVisibleChildTextFragment1.tabIndex)
         assertEquals(0, userVisibleChildTextFragment1.pageIndex)
 
-        fragmentScenario1.onFragment {
+        fragmentScenario1.onFragmentSleep100ms {
             userVisibleChildTextFragment1.parentFragment!!.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
         }
-        Thread.sleep(200)
         userVisibleChildTextFragment1 = fragment1.childFragmentManager.fragments.findUserVisibleFragmentRecursivelyByResumed() as TextFragment
         assertEquals(2, userVisibleChildTextFragment1.tabIndex)
         assertEquals(1, userVisibleChildTextFragment1.pageIndex)
@@ -480,10 +447,9 @@ class FragmentxTest {
             var childFragmentIndex2 = activity.findFragmentByViewPagerCurrentItem(2) as ResumeOnlyCurrentChildViewPagerFragment?
             assertNull(childFragmentIndex2)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 activity.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
 
             childFragmentIndex0 = activity.findFragmentByViewPagerCurrentItem(0) as ResumeOnlyCurrentChildViewPagerFragment?
             assertEquals(0, childFragmentIndex0!!.tabIndex)
@@ -494,10 +460,9 @@ class FragmentxTest {
             childFragmentIndex2 = activity.findFragmentByViewPagerCurrentItem(2) as ResumeOnlyCurrentChildViewPagerFragment?
             assertEquals(2, childFragmentIndex2!!.tabIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 activity.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
             }
-            Thread.sleep(200)
 
             childFragmentIndex0 = activity.supportFragmentManager.fragments.findFragmentByViewPagerCurrentItem(0) as ResumeOnlyCurrentChildViewPagerFragment?
             assertNull(childFragmentIndex0)
@@ -519,10 +484,9 @@ class FragmentxTest {
             var childTextFragment2 = childFragmentIndex2.findFragmentByViewPagerCurrentItem(2) as TextFragment?
             assertNull(childTextFragment2)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 childFragmentIndex2.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
 
             childTextFragment0 = childFragmentIndex2.findFragmentByViewPagerCurrentItem(0) as TextFragment?
             assertEquals(2, childTextFragment0!!.tabIndex)
@@ -536,10 +500,9 @@ class FragmentxTest {
             assertEquals(2, childTextFragment2!!.tabIndex)
             assertEquals(2, childTextFragment2.pageIndex)
 
-            activityScenario.onActivity {
+            activityScenario.onActivitySleep100ms {
                 childFragmentIndex2.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 2
             }
-            Thread.sleep(200)
 
             childTextFragment0 = childFragmentIndex2.findFragmentByViewPagerCurrentItem(0) as TextFragment?
             assertNull(childTextFragment0)
@@ -595,10 +558,9 @@ class FragmentxTest {
             assertFalse(grandsonFragment11.isResumed && grandsonFragment11.userVisibleHint)
             assertFalse(grandsonFragment11.isUserVisibleHintWithParent())
 
-            fragmentScenario.onFragment {
+            fragmentScenario.onFragmentSleep100ms {
                 fragment.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
 
             assertFalse(childFragment0.isResumed && childFragment0.userVisibleHint)
             assertFalse(childFragment0.isUserVisibleHintWithParent())
@@ -659,10 +621,9 @@ class FragmentxTest {
             assertFalse(grandsonFragment11.isResumed && grandsonFragment11.userVisibleHint)
             assertFalse(grandsonFragment11.isUserVisibleHintWithParent())
 
-            fragmentScenario.onFragment {
+            fragmentScenario.onFragmentSleep100ms {
                 fragment.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
 
             assertFalse(childFragment0.isResumed && childFragment0.userVisibleHint)
             assertFalse(childFragment0.isUserVisibleHintWithParent())
@@ -726,10 +687,9 @@ class FragmentxTest {
             assertFalse(grandsonFragment11.isResumed)
             assertFalse(grandsonFragment11.isResumedWithParent())
 
-            fragmentScenario.onFragment {
+            fragmentScenario.onFragmentSleep100ms {
                 fragment.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
 
             assertFalse(childFragment0.isResumed)
             assertFalse(childFragment0.isResumedWithParent())
@@ -790,10 +750,9 @@ class FragmentxTest {
             assertTrue(grandsonFragment11.isResumed)
             assertTrue(grandsonFragment11.isResumedWithParent())
 
-            fragmentScenario.onFragment {
+            fragmentScenario.onFragmentSleep100ms {
                 fragment.view!!.findViewById<ViewPager>(R.id.viewPagerAt_viewPager).currentItem = 1
             }
-            Thread.sleep(200)
 
             assertTrue(childFragment0.isResumed)
             assertTrue(childFragment0.isResumedWithParent())