+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * change: Fragmentx added isUserVisibleHintWithParent, isResumedWithParent, findUserVisibleFragmentRecursivelyByUserVisibleHint, findUserVisibleFragmentRecursivelyByResumed method
 * change: Fragmentx's findUserVisibleFragmentByUserVisibleHint, findUserVisibleFragmentByResumed methods no longer traverse grandchildren Fragment
 * new: Added FragmentInsider class, which contains getMaxState, getMaxStateWithParent methods
-* fix: Fix the bug of extra parameters of KClass<T>.launchFragment(KClass<T>) function of Testx.kt 
+* fix: Fix the bug of extra parameters of KClass<T>.launchFragment(KClass<T>) function of Testx.kt
+* new: Testx added onFragmentSleep, onFragmentSleep100ms, onFragmentSleep200ms, onActivitySleep, onActivitySleep100ms, onActivitySleep200ms methods
 
 ## v1.0.0-rc01